Manage the DB with cn=admin,dc=jeffmccune,dc=net
Without this patch the OpenLDAP server is being configured with the
example domain suffix.  This patch addresses the problem by changing the
domain database to be jeffmccune.net.

The server is be manageable using the DN
cn=admin,dc=jeffmccune,dc=net and the password Password1 and the port
forwarded on the host.  For example, from the vagrant host:

    ldapsearch -h localhost -p 10389 \
      -D cn=admin,dc=jeffmccune,dc=net \
      -w Password1 \
      -b dc=jeffmccune,dc=net
    # extended LDIF
    #
    # LDAPv3
    # base <dc=jeffmccune,dc=net> with scope subtree
    # filter: (objectclass=*)
    # requesting: ALL
    #
    # search result
    search: 2
    result: 32 No such object

The bind is successful but there are no objects in the database.