Load LDAP schema and base structure for accounts
Without this patch the LDAP database is empty.  This is a problem
because there needs to be schemas loaded and a basic structure in place
for user and group management.

This patch addresses the problem by automatically loading the schemas
for core, cosine, inetorgperson, and nis.  Once loaded, the basic
structure of the Groups and Users OU's are loaded.  Finally, a single
user account is created, uid=jeff with password Password1.

This should be enough data automatically loaded to serve accounts to
Jenkins.