Add site::jenkins class to configure the jenkins service
Without this patch we aren't configuring the Jenkins service, which is
necessary for the demo.  This patch addresses the problem by adding the
site::jenkins class to the sharedvm node which installs Jenkins and
configures it with some basic plugins.

No user creation or LDAP configuration is included in this patch, only
the basic starting of the service.