Bind Jenkins to LDAP
Without this patch jenkins has no authentication.  This is a problem
because we're not really demonstrating a distributed system.  This patch
addresses the problem by binding jenkins to LDAP for user
authentication.  With this patch applied Jenkins is configured to lookup
user accounts in LDAP and the user "jeff" with password "Password1" will
work to log in to the jenkins service.