Update ldap dockerfile to use centos7vps base image (REVISIT)
Without this patch the ldap Dockerfile does not build a suitable base
image for ldap because it starts from the centos:centos7 image which
contains a fake systemd package.

This patch addresses the problem by changing the ldap container build
file to use the centos7vps image which contains a functional systemd
implementation.

The build process is still incomplete because systemd is not running
during the docker build phase, however it will be running during the run
phase.

The puppet code still needs to be modified to account for systemd not
being present during build time.

REVISIT: With this change, does it still make sense to change the base
image to one that includes systemd?  It may make more sense to simply
override the Service[slapd] resource to use a start command that does
not rely on systemd.