Add `docker build` Dockerfile for ldap systemd container
Without this patch there is no single Dockerfile that implements a build
process that results in a working LDAP container.  This patch addresses
the problem by using the centos7vps base image, which includes systemd,
to build a LDAP service container.

The LDAP service can be started, mapping 389 on the host to 389 in the
container:

    CID=$(docker run -p 127.0.0.1:389:389 --privileged -d ldap_sysd)

This patch modifies the original Puppet code for the shared jenkins+ldap
VM overriding the Service[slapd] resource.  This is necessary because
Puppet defaults to systemd service management while systemd is not
running during the build phase.  The override manually starts the LDAP
process as a dependency of loading the initial schema and branches.

This approach differs from the approach of "side-loading" [1] the data
using two containers at runtime.

Details:

Start the service on port 389:

    $ CID=$(docker run -p 127.0.0.1:389:389 --privileged -d --name ldap ldap_sysd)

List the service:

    $ docker ps
    CONTAINER ID        IMAGE               COMMAND             CREATED             STATUS              PORTS                    NAMES
    8c2daa7ca724        ldap_sysd:latest    "/usr/sbin/init"    3 seconds ago       Up 3 seconds        127.0.0.1:389->389/tcp   ldap

There are 5 processes using the systemd approach:

    $ sudo ps auxww --forest
    root      1378  1.1  0.4 645540 39092 ?        Ssl  11:33   0:21 /usr/bin/docker -d --selinux-enabled
    root      8493  0.0  0.0 180136  4864 ?        Sl   12:04   0:00  \_ docker-proxy -proto tcp -host-ip 127.0.0.1 -host-port 389 -container-ip 172.17.0.10 -container-port 389
    root      8499  0.0  0.0  44616  2696 ?        Ss   12:04   0:00  \_ /usr/sbin/init
    root      8528  0.0  0.0  42992  3872 ?        Ss   12:04   0:00      \_ /usr/lib/systemd/systemd-journald
    55        8578  0.2  9.1 856940 734732 ?       Ssl  12:04   0:00      \_ /usr/sbin/slapd -u ldap -h ldapi:/// ldap:///

Access the service:

    ldapsearch -h localhost -p 389 \
               -D cn=admin,dc=jeffmccune,dc=net \
               -w Password1 \
               -b dc=jeffmccune,dc=net

[1]: http://jpetazzo.github.io/2014/06/23/docker-ssh-considered-evil/