Add Dockerfile for lightweight ldap with no systemd
Without this patch the only docker base image for ldap uses systemd.
This is a problem because it is not as lightweight as running ldap
without systemd and it goes against the principle of Docker [1], which
is to run as minimal of a container as possible.

This patch addresses the problem by taking the systemd enabled base
image as a starting point and replacing /usr/sbin/init with
/usr/sbin/slapd as the entry point of a running container.

The container no longer needs to be started in privileged mode.
Starting and stopping the container is also much faster without having
to go through systemd to start slapd.

Start the container:

    $ CID=$(docker run -p 127.0.0.1:10389:389 -d --name ldap2 ldap_lite)

List the containers:

    $ docker ps
    CONTAINER ID        IMAGE               COMMAND                CREATED             STATUS              PORTS                      NAMES
    0dce5b7994ab        ldap_lite:latest    "/usr/sbin/slapd -u    47 seconds ago      Up 47 seconds       127.0.0.1:10389->389/tcp   ldap2
    8c2daa7ca724        0d0bfe8aafcf        "/usr/sbin/init"       49 minutes ago      Up 48 minutes       127.0.0.1:389->389/tcp     ldap

There are fewer processes running (2 without systemd compared to 6 with
systemd).  This table shows the comparison between the systemd container on 389
and the slapd container on 1039.

    /usr/bin/docker -d --selinux-enabled
     \_ docker-proxy -proto tcp -host-ip 127.0.0.1 -host-port 389 -container-ip 172.17.0.10 -container-port 389
     \_ /usr/sbin/init
     |   \_ /usr/lib/systemd/systemd-journald
     |   \_ /usr/sbin/slapd -u ldap -h ldapi:/// ldap:///
     |   \_ /bin/dbus-daemon --system --address=systemd: --nofork --nopidfile --systemd-activation
     |   \_ /usr/lib/systemd/systemd-logind
     \_ docker-proxy -proto tcp -host-ip 127.0.0.1 -host-port 10389 -container-ip 172.17.0.21 -container-port 389
     \_ /usr/sbin/slapd -u ldap -h ldapi:/// ldap:/// -d 0

[1]: http://jpetazzo.github.io/2014/06/23/docker-ssh-considered-evil/