Add build and run tasks for a jenkins systemd container
Without this patch there is no jenkins container build procedure.  This
is a problem because we need to build a jenkins base image prior to
running the jenkins service in a container.  This patch addresses the
problem by creating a Dockerfile and rake tasks to build the jenkins
base image and run the image.

The build phase contains the expected error of not being able to manage
the Service[jenkins] resource:

    Error: Could not start Service[jenkins]: Execution of '/usr/bin/systemctl start jenkins' returned 1: Failed to get D-Bus connection: No connection to service manager.
    Wrapped exception:
    Execution of '/usr/bin/systemctl start jenkins' returned 1: Failed to get D-Bus connection: No connection to service manager.
    Error: /Stage[main]/Jenkins::Service/Service[jenkins]/ensure: change from stopped to running failed: Could not start Service[jenkins]: Execution of '/usr/bin/systemctl start jenkins' returned 1: Failed to get D-Bus connection: No connection to service manager.
    Warning: /Stage[main]/Jenkins/Anchor[jenkins::end]: Skipping because of failed dependencies

Despite the error, jenkins runs in the resulting systemd enabled
container using `rake run`.