Make Service[jenkins] work in `docker build` (systemd)
Without this patch puppet does not successfully complete during the
`docker build` phase of the jenkins container.  This is a problem
because there should be no configuration errors present in the puppet
results.

The root cause of the problem is Puppet trying to manage the
Service[jenkins] when systemd is not running and operational.

This patch addresses the problem by overriding Class[jenkins::service].
The original module is modified from upstream because it specifically
guards against modifying a private class and uses resource class
declaration syntax which is problematic with class overrides.

With this patch applied, running `rake run` will run the jenkins
container in VPS systemd mode, which requires privileged access.  The
container will link to the ldap container and configure jenkins to
access the LDAP server at `ldap://ldap:389`