Remove systemd at runtime in jenkins:lite image
Without this patch the jenkins container runs from the jenkins:systemd
base image where the jenkins service starts up from systemd and docker
starts systemd as the initial entrypoint.

This is a problem because the container must run in privileged mode and
takes longer than necessary to start and stop the container.  There are
additional processes running, like systemd, journald, and dbus.

This patch addresses the problem by executing jenkins directly and
skipping over systemd in the container.

With this patch applied the two running containers take a minimal number
of processes:

    /usr/bin/docker -d --selinux-enabled
     \_ /usr/sbin/slapd -u ldap -h ldapi:/// ldap:/// -d 0
     \_ docker-proxy -proto tcp -host-ip 0.0.0.0 -host-port 8080 -container-ip 172.17.0.57 -container-port 8080
     \_ java -DJENKINS_HOME=/var/lib/jenkins -jar /usr/lib/jenkins/jenkins.war